Revert "Changed build process to vite to shrink bundle size"
This reverts commit 37f2620409510622394e1fd83432991d8bb2d4d8.
