--- netlify.toml
@@ -1,2 +1,0 @@
-[build.environment]
-     NODE_VERSION = "20.12.0"
