+++ package.json
@@ -1,36 +1,27 @@
 {
     "name": "translation-helps-rcl",
-    "version": "3.6.0",
-    "main": "dist/index.cjs.js",
-    "module": "dist/index.es.js",
+    "version": "3.5.16",
+    "main": "dist/index.js",
     "homepage": "https://translation-helps-rcl.netlify.app/",
     "repository": "https://github.com/unfoldingWord/translation-helps-rcl.git",
     "author": "mannycolon <colonmanuel7@gmail.com>",
     "license": "MIT",
-    "exports": {
-        ".": {
-            "import": "./dist/index.es.js",
-            "require": "./dist/index.cjs.js"
-        }
-    },
-    "engines": {
-        "node": ">=18.18.0"
-    },
     "scripts": {
-        "start": "NODE_OPTIONS=--openssl-legacy-provider npx styleguidist server",
-        "build": "NODE_OPTIONS=--openssl-legacy-provider npx styleguidist build",
-        "build:vite": "vite build",
+        "start": "styleguidist server",
+        "build": "styleguidist build",
         "test": "jest",
         "lint": "eslint ./src",
         "lint:fix": "eslint ./src --fix",
         "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
-        "prepublishOnly": "rm -fr ./dist && vite build",
+        "prepublishOnly": "rm -rf ./dist && babel ./src --out-dir ./dist -s inline",
         "postpublish": "git tag v$npm_package_version && git push origin v$npm_package_version"
     },
     "files": [
         "dist/*"
     ],
     "peerDependencies": {
+        "@material-ui/core": "^4.x",
+        "@material-ui/lab": "^4.x",
         "@mui/icons-material": "^5.x",
         "@mui/lab": "^5.x",
         "@mui/material": "^5.x",
@@ -40,6 +31,7 @@
         "prop-types": "^15.x",
         "react": "^17.x || ^18.x ",
         "react-dom": "^17.x || ^18.x",
+        "react-icons": "^4.x",
         "scripture-resources-rcl": "^5.x"
     },
     "devDependencies": {
@@ -55,10 +47,8 @@
         "@mui/lab": "^5.0.0-alpha.89",
         "@mui/material": "^5.8.7",
         "@mui/styles": "^5.8.7",
-        "@vitejs/plugin-react": "^4.2.1",
         "babel-jest": "^27.3.1",
         "babel-loader": "^8.2.2",
-        "eslint": "^9.2.0",
         "gitea-react-toolkit": "^2.1.2",
         "jest": "^27.3.1",
         "markdown-translatable": "^2.1.1",
@@ -66,16 +56,17 @@
         "prop-types": "^15.7.2",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
+        "react-icons": "^4.x",
         "react-styleguidist": "^11.1.1",
         "scripture-resources-rcl": "^5.2.2",
         "tc-ui-toolkit": "^5.3.3",
-        "vite": "^5.2.11",
         "webpack": "4.44.0",
         "word-aligner": "^1.0.0"
     },
     "dependencies": {
         "@gwdevs/extensible-rcl": "^1.0.1",
         "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest",
+        "axios": "^0.27.2",
         "bible-reference-range": "^1.1.0",
         "dcs-branch-merger": "1.4.1",
         "deep-equal": "^2.0.5",
