+++ src/components/BranchMerger/components/ErrorDialog.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import Button from '@material-ui/core/Button';
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
+import LinkChip from './LinkChip';
+
+export default function ErrorDialog({ onClose, title, open, isLoading = false, content, link, linkTooltip = "" }) {
+
+  return (
+    <Dialog open={open} onClose={onClose} aria-labelledby="form-dialog-title">
+      <DialogTitle id="form-dialog-title">{title}</DialogTitle>
+      <DialogContent>
+        <DialogContentText>
+          {content}
+        </DialogContentText>
+        {link ? <LinkChip link={link} linkTooltip={linkTooltip} /> : null}
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onClose} color="primary" disabled={isLoading}>
+          Accept
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
