--- src/components/BranchMerger/components/UpdateBranchButton.jsx
@@ -1,54 +1,0 @@
-import React from 'react'
-import {
-  Update as UpdateIcon,
-  UpdateDisabled as UpdateDisabledIcon,
-} from '@mui/icons-material'
-import { IconButton, Badge, CircularProgress, Tooltip } from '@mui/material'
-
-export default function UpdateBranchButton({
-  onClick = () => undefined,
-  isLoading = false,
-  blocked = false,
-  pending = false,
-  loadingProps,
-  title = 'Update my content',
-  ...props
-}) {
-  const Icon = blocked ? UpdateDisabledIcon : UpdateIcon
-  return (
-    <Tooltip title={title}>
-      <span>
-        <IconButton
-          key='update-from-master'
-          onClick={onClick}
-          aria-label={title}
-          style={{
-            cursor: 'pointer',
-            ...props.style,
-            ...(!pending | blocked ? { color: 'rgba(0, 0, 0, 0.26)' } : {}),
-          }}
-          // disabled={!pending | blocked}
-        >
-          {isLoading ? (
-            <CircularProgress
-              size={18}
-              style={{
-                margin: '3px 0px',
-              }}
-              {...loadingProps}
-            />
-          ) : (
-            <Badge
-              color='error'
-              variant='dot'
-              overlap='circular'
-              invisible={!pending}
-            >
-              <Icon id='update-from-master-icon' />
-            </Badge>
-          )}
-        </IconButton>
-      </span>
-    </Tooltip>
-  )
-}
