+++ src/components/Card/Card.js
@@ -1,9 +1,13 @@
 import React, { useState } from 'react'
+import { styled } from '@mui/material/styles';
 import PropTypes from 'prop-types'
 import styledComponents from 'styled-components'
-import { styled, Badge, IconButton } from '@mui/material'
 import {
-  Close as CloseIcon,
+  Badge,
+  IconButton
+} from '@mui/material'
+import CloseIcon from '@mui/icons-material/Close'
+import {
   MoreVert as MoreVertIcon,
   Announcement as AnnouncementIcon,
   DragIndicator as DragIndicatorIcon,
@@ -21,7 +25,7 @@ import SettingsCard from '../SettingsCard'
 import Scrollable from '../Scrollable'
 import { Extensible } from '@gwdevs/extensible-rcl'
 
-const PREFIX = 'Card'
+const PREFIX = 'Card';
 
 const classes = {
   title: `${PREFIX}-title`,
@@ -29,8 +33,8 @@ const classes = {
   chevronIcon: `${PREFIX}-chevronIcon`,
   pointerIcon: `${PREFIX}-pointerIcon`,
   children: `${PREFIX}-children`,
-  paddingRight: `${PREFIX}-paddingRight`,
-}
+  paddingRight: `${PREFIX}-paddingRight`
+};
 
 const StyledPaper = styled(Paper)(() => ({
   [`& .${classes.title}`]: {
@@ -67,8 +71,8 @@ const StyledPaper = styled(Paper)(() => ({
 
   [`& .${classes.paddingRight}`]: {
     padding: '0px 15px 0px 0px',
-  },
-}))
+  }
+}));
 
 const FlexDiv = styledComponents.div`
   display: flex;
@@ -101,9 +105,8 @@ const Navigation = ({
       id={`${baseId}_prev`}
       onClick={onPrevItem}
     />
-    <FlexDiv id={`${baseId}_nav`}>{`${itemIndex + 1} of ${
-      items.length
-    }`}</FlexDiv>
+    <FlexDiv id={`${baseId}_nav`}>{`${itemIndex + 1} of ${items.length
+      }`}</FlexDiv>
     <ChevronRightIcon
       className={classes.chevronIcon}
       id={`${baseId}_next`}
@@ -151,6 +154,7 @@ const Card = ({
   classes: { root, dragIndicator, header, children: childrenClassName },
 }) => {
   const [showMenu, setShowMenu] = useState(false)
+
 
   let settingsTitle = settingsTitle_
   if (!settingsTitle) {
@@ -353,7 +357,7 @@ const Card = ({
         enableAutoScrollToTop={enableAutoScrollToTop}
       />
     </StyledPaper>
-  )
+  );
 }
 
 Card.defaultProps = {
