+++ src/components/DraggableModal/DraggableModal.js
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react'
 import PropTypes from 'prop-types'
-import { Dialog } from '@mui/material'
+import Dialog from '@mui/material/Dialog'
 import Draggable from 'react-draggable'
 
 export default function DraggableModal({
@@ -11,30 +11,28 @@ export default function DraggableModal({
   onStartDrag,
   bounds,
   initialPosition,
-}) {
-  const PaperComponent = useCallback(
-    props => {
-      function onStart(e) {
-        if (onStartDrag) {
-          return onStartDrag(e)
-        }
-        return true
-      }
+ }) {
 
-      return (
-        <Draggable
-          handle='.draggable-dialog-title'
-          cancel={'[class*="MuiDialogContent-root"]'}
-          onStart={onStart}
-          bounds={bounds}
-          defaultPosition={initialPosition}
-        >
-          <div {...props} />
-        </Draggable>
-      )
-    },
-    [onStartDrag, bounds, initialPosition]
-  )
+  const PaperComponent = useCallback( (props) => {
+    function onStart(e) {
+      if (onStartDrag) {
+        return onStartDrag(e)
+      }
+      return true
+    }
+
+    return (
+      <Draggable
+        handle='.draggable-dialog-title'
+        cancel={'[class*="MuiDialogContent-root"]'}
+        onStart={onStart}
+        bounds={bounds}
+        defaultPosition={initialPosition}
+      >
+        <div {...props} />
+      </Draggable>
+    )
+  }, [onStartDrag, bounds, initialPosition])
 
   return (
     <Dialog
