+++ src/components/SettingsCard/SettingsCard.js
@@ -1,20 +1,18 @@
 import React, { Fragment } from 'react'
+import { styled } from '@mui/material/styles';
 import PropTypes from 'prop-types'
-import {
-  styled,
-  Switch,
-  Divider,
-  Checkbox,
-  FormGroup,
-  Typography,
-  FormControlLabel,
-} from '@mui/material'
+import Switch from '@mui/material/Switch'
+import Divider from '@mui/material/Divider'
+import Checkbox from '@mui/material/Checkbox'
+import FormGroup from '@mui/material/FormGroup'
+import Typography from '@mui/material/Typography'
+import FormControlLabel from '@mui/material/FormControlLabel'
 import { Extensible } from '@gwdevs/extensible-rcl'
 import DraggableModal from '../DraggableModal'
 import FontSizeSlider from '../FontSizeSlider'
 import Card from '../Card'
 
-const PREFIX = 'SettingsCard'
+const PREFIX = 'SettingsCard';
 
 const classes = {
   switchBase: `${PREFIX}-switchBase`,
@@ -34,10 +32,14 @@ const classes = {
   checkboxes: `${PREFIX}-checkboxes`,
   cardRemovalSection: `${PREFIX}-cardRemovalSection`,
   removeText: `${PREFIX}-removeText`,
-  formControl: `${PREFIX}-formControl`,
-}
+  formControl: `${PREFIX}-formControl`
+};
 
-const StyledDraggableModal = styled(DraggableModal)(({ theme }) => ({
+const StyledDraggableModal = styled(DraggableModal)((
+  {
+    theme
+  }
+) => ({
   [`& .${classes.root2}`]: {
     width: '100%',
   },
@@ -97,12 +99,12 @@ const StyledDraggableModal = styled(DraggableModal)(({ theme }) => ({
   [`& .${classes.formControl}`]: {
     margin: theme.spacing(1),
     minWidth: 120,
-  },
-}))
+  }
+}));
 
 const BlueSwitch = Switch
 
-const BlueCheckbox = props => <Checkbox color='default' {...props} />
+const BlueCheckbox = (props => <Checkbox color='default' {...props} />)
 
 const SettingsCard = ({
   open,
@@ -121,6 +123,8 @@ const SettingsCard = ({
   hideMarkdownToggle,
   getCustomComponent,
 }) => {
+
+
   const handleCheckboxClick = event => {
     event.preventDefault()
     let newFilters = []
@@ -166,9 +170,8 @@ const SettingsCard = ({
                     classes={{
                       switchBase: classes.switchBase,
                       checked: classes.checked,
-                      track: classes.track,
-                    }}
-                  />
+                      track: classes.track
+                    }} />
                 }
                 classes={{ label: classes.switchLabel }}
                 label='Markdown View'
@@ -208,9 +211,8 @@ const SettingsCard = ({
                             checked={filters.includes(header)}
                             classes={{
                               root: classes.root,
-                              checked: classes.checked2,
-                            }}
-                          />
+                              checked: classes.checked2
+                            }} />
                         }
                       />
                     ))}
@@ -232,7 +234,7 @@ const SettingsCard = ({
         </Extensible>
       </Card>
     </StyledDraggableModal>
-  )
+  );
 }
 
 SettingsCard.defaultProps = {
