+++ src/components/TqContent/TqContent.js
